+++ src/Eccube/Repository/CustomerRepository.php
@@ -47,7 +47,7 @@ use Symfony\Component\Security\Core\User\UserProviderInterface;
  *
  * @Repository
  */
-class CustomerRepository extends AbstractRepository implements UserProviderInterface
+class CustomerRepository extends AbstractRepository
 {
     /**
      * @Inject("eccube.queries")
@@ -95,69 +95,6 @@ class CustomerRepository extends AbstractRepository implements UserProviderInter
             ->setStatus($CustomerStatus);
 
         return $Customer;
-    }
-
-    /**
-     * Loads the user for the given username.
-     *
-     * This method must throw UsernameNotFoundException if the user is not
-     * found.
-     *
-     * @param string $username The username
-     *
-     * @return UserInterface
-     *
-     * @see UsernameNotFoundException
-     *
-     * @throws UsernameNotFoundException if the user is not found
-     */
-    public function loadUserByUsername($username)
-    {
-        $Customer = $this->findOneBy([
-            'email' => $username,
-            'Status' => CustomerStatus::REGULAR,
-        ]);
-
-        if (is_null($Customer)) {
-            throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
-        }
-
-        return $Customer;
-    }
-
-    /**
-     * Refreshes the user for the account interface.
-     *
-     * It is up to the implementation to decide if the user data should be
-     * totally reloaded (e.g. from the database), or if the UserInterface
-     * object can just be merged into some internal array of users / identity
-     * map.
-     *
-     * @param UserInterface $user
-     *
-     * @return UserInterface
-     *
-     * @throws UnsupportedUserException if the account is not supported
-     */
-    public function refreshUser(UserInterface $user)
-    {
-        if (!$user instanceof Customer) {
-            throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', get_class($user)));
-        }
-
-        return $this->loadUserByUsername($user->getUsername());
-    }
-
-    /**
-     * Whether this provider supports the given user class.
-     *
-     * @param string $class
-     *
-     * @return bool
-     */
-    public function supportsClass($class)
-    {
-        return $class === Customer::class;
     }
 
     public function getQueryBuilderBySearchData($searchData)
