+++ src/Eccube/Security/Core/User/CustomerProvider.php
@@ -1,0 +1,83 @@
+<?php
+
+namespace Eccube\Security\Core\User;
+
+use Eccube\Entity\Customer;
+use Eccube\Entity\Master\CustomerStatus;
+use Eccube\Repository\CustomerRepository;
+use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
+use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
+
+class CustomerProvider implements UserProviderInterface
+{
+    /**
+     * @var CustomerRepository
+     */
+    protected $customerRepository;
+
+    public function __construct(CustomerRepository $customerRepository)
+    {
+        $this->customerRepository = $customerRepository;
+    }
+
+    /**
+     * Loads the user for the given username.
+     *
+     * This method must throw UsernameNotFoundException if the user is not
+     * found.
+     *
+     * @param string $username The username
+     *
+     * @return UserInterface
+     *
+     * @throws UsernameNotFoundException if the user is not found
+     */
+    public function loadUserByUsername($username)
+    {
+        $Customer = $this->customerRepository->findOneBy([
+            'email' => $username,
+            'Status' => CustomerStatus::REGULAR,
+        ]);
+
+        if (null === $Customer) {
+            throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
+        }
+
+        return $Customer;
+    }
+
+    /**
+     * Refreshes the user.
+     *
+     * It is up to the implementation to decide if the user data should be
+     * totally reloaded (e.g. from the database), or if the UserInterface
+     * object can just be merged into some internal array of users / identity
+     * map.
+     *
+     * @return UserInterface
+     *
+     * @throws UnsupportedUserException if the user is not supported
+     */
+    public function refreshUser(UserInterface $user)
+    {
+        if (!$user instanceof Customer) {
+            throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', get_class($user)));
+        }
+
+        return $this->loadUserByUsername($user->getUsername());
+    }
+
+    /**
+     * Whether this provider supports the given user class.
+     *
+     * @param string $class
+     *
+     * @return bool
+     */
+    public function supportsClass($class)
+    {
+        return Customer::class === $class;
+    }
+}
